Fix test for all python version
Error was

FutureWarning: The input object of type 'ultrasound_rawdata_exchange.bindings.UrxVector3D' is an array-like implementing one of the corresponding protocols (`__array__`, `__array_interface__` or `__array_struct__`); but not a sequence (or 0-D). In the future, this object will be coerced as if it was first converted using `np.array(obj)`. To retain the old behaviour, you have to either modify the type 'ultrasound_rawdata_exchange.bindings.UrxVector3D', or assign to an empty array created with `np.empty(correct_shape, dtype=object)`.

ValueError: setting an array element with a sequence. The requested array has an inhomogeneous shape after 1 dimensions. The detected shape was (2,) + inhomogeneous part.